More self code review
